src/World/Entity/Objects/Characters/player.cpp
Collision rect size now is independent of animation frame size.

I was supposed to deal with collision rect offset, but I decided that it would complicate many things in physics engine logic. So for now top and left bound positions are dependent of texture. So we have to make our textures not have an space between left bound of the texture to left bound of the actual character graphic. The same thing with top bounds.
Maybe we'll provide the collision rect offset support after the whole physics engine is made.