Refactor and fixing issues in Physics.
In PhysicsEngine::handleKinematicCollisions there was an issue. For example it was checking is player colliding with player. Ofcourse player always would collide with player. So it calculated this collision further and it was obvious waste of time. This issue was fixed.
I also fixed some formatting.
And I added some logs which will be very helpful in debugging kinematic collisions.