After big changes in physics module physics still doesn't work, but I fixed some things.
The camera now is working correctly.

Scene objects like zombie now have proper positions.

Probably the collision bodies were checking positions with themselfes which resulted in nan value which was returned by this method KinematicCollisionHandler::getDirectionOfPush().
The hypotenuse was calculated. Of course it was 0. Later something was devided by hypotenuse which resulted in nan value. I temporarily fixed it with simple if.

There was a problem with refrences to CollisionBody objects which are hold by Character. The references were pointing to elements of the vector. And when the vector was expanded the references didn't point to anything specific. For now I called reserve method on these vectors. reserve(1000) for kinematic bodies and reserve(10000) for static bodies.

I already fixed a lot, but still physics is not working at all. Neither static nor dynamic collisions.