Add minimal delta time constrain
For now minimal delta time constrain is 1/20 sec which means that when
we press button to move our character we go at the same speed in real
time if game framerate is between 20 FPS and 60 FPS. But if framerate is
dropping below the 20 FPS we pass 1/20 sec all the time so our character
is moving slower the usually.
I had to do this constrain because otherwise physics module could do
weird things which could affect in crash.