Make zombies walk in main menu
This required some changes in AIManager because zombies were trying to
find the path to player but there is no player on main menu scene. So
now TiledGameObjectParser sends the information is player on scene. When
he's not on scene zombies don't try to look for him.
Also I have to add try catch block in one method in Zombie because
exxception was thrown.