Bump scene tree rendering architecture
The main reason of this change is desynchronization between scene tree
and renderer. Because of that managing and rendering game objects was
quite hard, besides now we don't have to store all those pointers to game
objects in renderer layers so we don't waste so much memory.

(on this commit there are bugs)