Cut input() in game objects, scenes and efficiency register
This was done to have better performance and design of input handling
Before when we handled all kind of input in input() function we had to
store boolean values in order to use them in update() function.
Now we don't waste all this memory for storing bools.
Immediate input should be handled in update() functions.
But events should call callback function; for now I used temporary
solution to handle them, but I'll do callbacks in next commit.