Bump event handling
Before this commit events were handled through abstraction in Input module. The
reason of this design was making event handling syntax similar to syntax
of immediate input. That proved unconvinient and bad for several reasons.
That's why I had to change it.
So right now events handling starts in Game::handleEvents() and events
are send down to several modules and handled in (ClassName)::handleEvent() functions.
Ofcourse I cut useless classes in Input module.