Cut events directory and replace ph::Event with sf::Events everywhere
The whole pophead events system was designed to support key rebindings.
Recently I made game design decision that we won't support key
rebindings. So stuff in Events dir has become useless. That didn't
bother me that much but now I want to make joystick support and I tried
to add joysticks to previous events system and it would be pain in the
ass and it would create a lot of unnecessary complexity. So I just decided
to delete events system.